requirements.txt
django==4.1.7
pyyaml==6.0.0
gunicorn==20.1.0
uvicorn==0.21.1
django-activity-stream==1.4.2
pillow==9.5.0
psycopg2-binary==2.9.6
django-allauth==0.54.0
djangorestframework==3.14.0
python-dotenv==1.0.0
django-reversion==5.0.4